Add Dash conference
Dash is an annual conference about building and scaling the next generation of applications, infrastructure, and technical teams. The two-day conference brings together more than 1,200 engineers across dev and ops who are taking their systems and organizations to the next level of velocity, performance, reliability, and scale. Dash features a mix of speaking sessions, hands-on labs and trainings, and opportunities to make new connections and exchange ideas.